re-structed files and folders
