--- src/entity.js
@@ -1,7 +1,0 @@
-import movement from './movementStatemachine';
-
-export const Entity = {
-  player: {
-    movement,
-  },
-}
