--- src/gamepadEventListener.js
@@ -1,5 +1,0 @@
-// returns a singleton for listening to gamepad events
-
-const listeners = {};
-const getGamepadEventsListener = () => {
-}
