+++ src/getLogger/getDiff.js
