+++ src/getLogger/index.js
@@ -1,30 +1,28 @@
-import { updateProps } from '../entities';
-import { update } from '../lastUpdated';
 import { printType } from './printers';
 import { getDiff } from './getDiff';
+import { getFilter } from './getFilter';
 
 // returns middleware and a function to change the filters
 // This is an opinionated logger for my use for this game
 const logger = () => {
   let filters = {};
-  const filter = ({ action }) =>
-    // TODO: use type insteadof toString??
-    (filters.includeUpdate || action.type !== updateProps.toString())
-    && (filters.includeLastUpdated || action.type !== update.toString());
+  let oldCount = -1;
 
   return {
+    changeFilters: _filters => {
+      filters = _filters;
+      oldCount = -1;  // This gaurentees we'll re-render the logs next action.
+    },
     middleware: store => {
       const logs = [];
-      let oldCount = logs.length;
       let oldResult = store.getState();
       return next => action => {
         const result = next(action);
-        // TODO: maybe clone the result andold result before putting it in getDiff or maybe just copy and freeze?? I dunno...
         const diff = getDiff(store.getState(), oldResult);
         oldResult = store.getState();
         logs.unshift({ action, diff, time: Date.now() });
-        const filteredLogs = logs.filter(filter);
-        // TODO: also do this hen filters change. Consider my above restriction on updating logger
+        const filteredLogs = logs.filter(getFilter(filters));
+        // we check lazily if the filtered logs has changed.
         if (oldCount !== filteredLogs.length) {
           printType(filteredLogs);
         }
@@ -32,7 +30,6 @@ const logger = () => {
         return result;
       };
     },
-    changeFilters: _filters => filters = _filters,
   };
 };
