+++ src/index.js
@@ -2,41 +2,14 @@ import 'regenerator-runtime/runtime.js';
 import 'end-polyFills';
 
 import { createStore, applyMiddleware, compose } from 'redux';
-import { createYieldEffectMiddleware } from 'redux-yield-effect';
-// import { put, fork, join } from 'redux-yield-effect/lib/effects';
-import { tickMiddleware, resumeTicks, /* pauseTicks */ } from 'effect-tick';
-import { createEntity } from './entities/index.js';
+import { createEntity } from './entities';
 import reducer from './reducer.js';
-import metaSelector from 'redux-meta-selector';
-import loadResources from './loadResources';
-import graphics from './graphics.js';
-import { setControls } from './Controls';
-import { typeDefinitions } from './entities/typeDefinitions';
-import { update } from './lastUpdated';
-import getLogger from './logger'; // TODO: rename file to getLogger
-
-// TODO: clean up this index file a lot!
+import { replaceAllTheModules, setUpdate, loadRaf } from './bootstrap';
+import { setControls } from './controls';
+import getLogger from './getLogger';
 
 document.addEventListener('DOMContentLoaded', firstLoad);
 
-window.raf    // the raf to cancel on HMR
-window.store  // the store to update on HMR
-
-const step = (ctx, resources) => dt => {
-  graphics(ctx, window.store.getState(), resources, dt);
-  window.raf = window.requestAnimationFrame(step(ctx, resources));
-};
-
-const loadRaf = () => {
-  loadResources().then(resources =>  {
-    const ctx = document.querySelector('canvas').getContext('2d');
-    if (window.raf) window.cancelAnimationFrame(window.raf);
-    window.raf = window.requestAnimationFrame(step(ctx, resources));
-  });
-};
-
-//TODO: spriteWidth should be defined elsewhere.. ...
-const spriteWidth = 96;
 async function firstLoad() {
   const logger = getLogger();
   const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
@@ -44,22 +17,14 @@ async function firstLoad() {
     reducer,
     composeEnhancers(applyMiddleware(
       logger.middleware,
-      /*
-      createYieldEffectMiddleware(),
-      tickMiddleware,
-      metaSelector,
-      */
     )),
   );
 
   // load request animation frame
   loadRaf();
 
-  // start ticks
-  window.store.dispatch(resumeTicks());
-
-
   // create the first player entity
+  const spriteWidth = 96;
   window.store.dispatch(createEntity({
     props: {
       x: spriteWidth,
@@ -77,29 +42,7 @@ async function firstLoad() {
   window.updateInterval = setUpdate();
 }
 
-const setUpdate = () => setInterval(() => {
-  const { entities, lastUpdated } = window.store.getState();
-  const dt = (Date.now() - lastUpdated) / 1000;
-  Object.values(entities).forEach(entity =>
-    typeDefinitions[entity.type].update(entity, dt, window.store.dispatch));
-  // call cleanup
-  window.store.dispatch(update(Date.now()));
-}, 500);
-
-
 if(module.hot) {
-  // Called after every change.
-  // Use window globals to keep track of old state.
   module.hot.accept();
-  if (window.raf) loadRaf();
-  if (window.store) window.store.replaceReducer(reducer);
-  if (window.controlsInterval) {
-    clearInterval(window.controlsInterval);
-    window.controlsInterval = setControls(window.store.dispatch);
-  }
-  if (window.updateInterval) {
-    clearInterval(window.updateInterval);
-    window.updateInterval = setUpdate();
-  }
-  // NOTE: I might have to add controller in here
+  replaceAllTheModules();
 }
