+++ thingstoconsider.rtf
@@ -59,7 +59,7 @@ feat/debugger (redux devtools is too buggy, overwrites actions and shows undefin
 chore/running-reorganize\
 	*- is HMR working for the update cycle???\
  	*- test HMR with changes to reducer code, like like vx vy changes based on input.\
-	- organize files and folders before testing, for sanity sake\
+	- organize files and folders before testing\
 \
 extra\
 	- fix auto indenting in vim\
